comparaciones

+++ Adriano/spec/008-paseNumeros-spec.js
@@ -4,9 +4,11 @@ describe("Manejo de parse", function(){
 
   var cadena1=5;
   var cadena2=7;
+
   var numero1=parseInt(cadena1);
   var numero2=parseInt(cadena2);
   var suma=numero1+numero2;
+
   expect(suma).toBe(12);
 
   });
@@ -14,10 +16,12 @@ describe("Manejo de parse", function(){
 
   var cadena1=5.2;
   var cadena2=7.3;
-  var numero1=parseInt(cadena1);
-  var numero2=parseInt(cadena2);
+
+  var numero1=parseFloat(cadena1);
+  var numero2=parseFloat(cadena2);
   var suma=numero1+numero2;
+
   expect(suma).toBe(12.5);
 
-  });
+});
 });
